requirements.txt
boto3
pprint
requests
requests_oauthlib
python-twitter
